requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
alembic
requests
pandas
scikit-learn
numpy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
beautifulsoup4
lxml
textblob
schedule
nltk
apscheduler
google-genai
pydantic-settings
